+++ interfaces/web/advanced_controllers/__init__.py
@@ -3,3 +3,5 @@ from flask import Blueprint
 advanced = Blueprint('advanced', __name__, url_prefix='/advanced', template_folder="../advanced_templates")
 
 from . import home
+from . import configuration
+from . import strategy_optimizer