+++ interfaces/web/controllers/configuration.py
@@ -2,11 +2,11 @@ from flask import render_template, request, jsonify
 
 from config.cst import CONFIG_EXCHANGES, CONFIG_CATEGORY_SERVICES, CONFIG_CATEGORY_NOTIFICATION, CONFIG_TRADING, \
     CONFIG_TRADER, CONFIG_SIMULATOR, CONFIG_CRYPTO_CURRENCIES, GLOBAL_CONFIG_KEY, EVALUATOR_CONFIG_KEY, \
-    CONFIG_TRADER_REFERENCE_MARKET
+    CONFIG_TRADER_REFERENCE_MARKET, TRADING_CONFIG_KEY
 from interfaces.web import server_instance
-from interfaces.web.models.configuration import get_evaluator_config, update_evaluator_config, \
+from interfaces.web.models.configuration import get_strategy_config, update_evaluator_config, \
     get_evaluator_startup_config, get_services_list, get_symbol_list, update_global_config, get_all_symbol_list, \
-    get_full_exchange_list, get_edited_config
+    get_full_exchange_list, get_edited_config, update_trading_config, get_trading_startup_config
 from interfaces.web.util.flask_util import get_rest_reply
 
 
@@ -23,6 +23,10 @@ def config():
             if GLOBAL_CONFIG_KEY in request_data and request_data[GLOBAL_CONFIG_KEY]:
                 success = update_global_config(request_data[GLOBAL_CONFIG_KEY])
 
+            # update trading config if required
+            if TRADING_CONFIG_KEY in request_data and request_data[TRADING_CONFIG_KEY]:
+                success = success and update_trading_config(request_data[TRADING_CONFIG_KEY])
+
             # update evaluator config if required
             if EVALUATOR_CONFIG_KEY in request_data and request_data[EVALUATOR_CONFIG_KEY]:
                 success = success and update_evaluator_config(request_data[EVALUATOR_CONFIG_KEY])
@@ -34,6 +38,7 @@ def config():
 
             response = {
                 "evaluator_updated_config": request_data[EVALUATOR_CONFIG_KEY],
+                "trading_updated_config": request_data[TRADING_CONFIG_KEY],
                 "global_updated_config": request_data[GLOBAL_CONFIG_KEY],
                 removed_elements_key: request_data[removed_elements_key]
             }
@@ -62,10 +67,12 @@ def config():
                                ccxt_exchanges=sorted(get_full_exchange_list(True)),
                                services_list=service_list,
                                service_name_list=service_name_list,
-                               symbol_list=sorted(get_symbol_list([exchange for exchange in display_config[CONFIG_EXCHANGES]])),
+                               symbol_list=sorted(get_symbol_list([exchange
+                                                                   for exchange in display_config[CONFIG_EXCHANGES]])),
                                full_symbol_list=get_all_symbol_list(),
-                               evaluator_config=get_evaluator_config(),
-                               evaluator_startup_config=get_evaluator_startup_config()
+                               strategy_config=get_strategy_config(),
+                               evaluator_startup_config=get_evaluator_startup_config(),
+                               trading_startup_config=get_trading_startup_config()
                                )