+++ interfaces/web/static/js/components/configuration.js
@@ -264,13 +264,78 @@ function handle_save_buttons_success_callback(updated_data, update_url, dom_root
     create_alert("success", "Configuration successfully updated. Restart OctoBot for it to be applied", "");
 }
 
+function other_trading_mode_activated(){
+    let other_activated_modes_count = $("#trading-modes-config-root").children(".list-group-item-success").length;
+    return other_activated_modes_count > 1;
+}
+
+function deactivate_other_trading_modes(element) {
+    const element_id = element.attr("id");
+    $("#trading-modes-config-root").children(".list-group-item-success").each(function () {
+        const element = $(this);
+        if(element.attr("id") !== element_id){
+            element.attr(current_value_attr, "false");
+            update_element_temporary_look(element);
+        }
+    })
+}
+
+function update_requirement_activation(element) {
+    const required_elements = element.attr("requirements").split("'");
+    $("#evaluator-config-root").children(".config-element").each(function () {
+        const element = $(this);
+        if(required_elements.indexOf(element.attr("id")) !== -1){
+            element.attr(current_value_attr, "true");
+            update_element_temporary_look(element);
+            update_element_required_marker_and_usability(element, true);
+        }else{
+            element.attr(current_value_attr, "false");
+            update_element_temporary_look(element);
+            update_element_required_marker_and_usability(element, false);
+        }
+    });
+}
+
+function check_evaluator_configuration() {
+    const activated_trading_modes = $("#trading-modes-config-root").children(".list-group-item-success");
+    if(activated_trading_modes.length > 0){
+        const activated_trading_mode = activated_trading_modes;
+        const required_elements = activated_trading_mode.attr("requirements").split("'");
+        let at_lest_one_activated_element = false;
+        $("#evaluator-config-root").children(".config-element").each(function () {
+            const element = $(this);
+            if(required_elements.indexOf(element.attr("id")) !== -1) {
+                if (element.attr(current_value_attr).toLowerCase() === "true") {
+                    at_lest_one_activated_element = true;
+                }
+            }else{
+                update_element_required_marker_and_usability(element, false);
+            }
+        });
+        if(!at_lest_one_activated_element){
+            create_alert("error", "Trading modes require at least one strategy to work properly, please activate the " +
+                "strategy(ies) you want for the selected mode.", "");
+        }
+    }else{
+        create_alert("error", "No trading mode activated, OctoBot need at least one trading mode.", "");
+    }
+}
+
 function handle_evaluator_configuration_editor(){
-    $(".config-element").click(function(){
+    $(".config-element").click(function(e){
+        if (isDefined($(e.target).attr(no_activation_click_attr))){
+            // do not trigger when click on items with no_activation_click_attr set
+            return;
+        }
         const element = $(this);
 
-        if (element.hasClass(config_element_class)){
+        if (element.hasClass(config_element_class) && ! element.hasClass(disabled_class)){
 
-            if (element[0].hasAttribute(config_type_attr) && element.attr(config_type_attr) === evaluator_config_type){
+            if (element[0].hasAttribute(config_type_attr)
+                && (element.attr(config_type_attr) === evaluator_config_type
+                    || element.attr(config_type_attr) === trading_config_type)){
+
+                const is_trading_mode = element.attr(config_type_attr) === trading_config_type;
 
                 // build data update
                 let new_value = parse_new_value(element);
@@ -283,11 +348,20 @@ function handle_evaluator_configuration_editor(){
                     current_value = element.attr(current_value_attr);
                 }
 
-                // todo
                 if (current_value === "true"){
+                    if(is_trading_mode && !other_trading_mode_activated()){
+                        create_alert("error", "Impossible to disable all trading modes.", "");
+                        return;
+                    }
                     new_value = "false";
                 }else if(current_value === "false"){
                     new_value = "true";
+                    if(is_trading_mode){
+                        deactivate_other_trading_modes(element);
+                    }
+                }
+                if(is_trading_mode){
+                    update_requirement_activation(element);
                 }
 
                 // update current value
@@ -330,4 +404,6 @@ $(document).ready(function() {
     register_edit_events();
 
     register_exit_confirm_function(something_is_unsaved);
+
+    check_evaluator_configuration();
 });