+++ tools/number_util.py
@@ -1,0 +1,2 @@
+def round_into_str_with_max_digits(number, digits_count):
+    return "{:.{}f}".format(round(number, digits_count), digits_count)